+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## Changelog
 
 ### Unreleased
-- Improvements:
+- Features:
     - Implement clickable image through CustomJson [RMCCX-7233]
+- Fixes:
+    - Fix clickable image behaviour for invalid url [RMCCX-7351]
 
 ### 8.4.0 (2024-09-17) 
 - Features: